requirements.txt
pillow
fastapi
uvicorn
pytest
mock
starlette
jinja2
pytest